+++ lib/specl/sandbox.lua
@@ -220,7 +220,8 @@ end
 -- correctly.
 local function inner_closures (env, state)
   env.load = function (ld, source, _, fenv)
-    local fn = load (ld, source)
+    local fn, err = load (ld, source)
+    if fn == nil then return nil, err end
     return function ()
       setfenv (fn, fenv or env)
       return fn ()
@@ -228,7 +229,8 @@ local function inner_closures (env, state)
   end
 
   env.loadfile = function (filename, _, fenv)
-    local fn = loadfile (filename)
+    local fn, err = loadfile (filename)
+    if fn == nil then return nil, err end
     return function ()
       setfenv (fn, fenv or env)
       return fn ()