+++ lib/specl/formatter/tap.lua
@@ -1,5 +1,9 @@
+--[[
+ Behaviour Driven Development for Lua 5.1, 5.2 & 5.3.
+ Copyright (C) 2013-2018 Francois Perrad
+]]
+
 -- Test Anything Protocol style formatter.
---
 
 local util = require "specl.util"