+++ lib/specl/shell.lua
@@ -1,20 +1,7 @@
--- Shell and file helpers.
--- Written by Gary V. Vaughan, 2013
---
--- Copyright (c) 2013-2017 Gary V. Vaughan
---
--- Specl is free software distributed under the terms of the MIT license;
--- it may be used for any purpose, including commercial purposes, at
--- absolutely no cost without having to ask permission.
---
--- The only requirement is that if you do use Specl, then you should give
--- credit by including the appropriate copyright notice somewhere in your
--- product or its documentation.
---
--- You should have received a copy of the MIT license along with this
--- program; see the file LICENSE.md.  If not, a copy can be downloaded
--- from <https://mit-license.org>.
-
+--[[
+ Behaviour Driven Development for Lua 5.1, 5.2 & 5.3.
+ Copyright (C) 2013-2018 Gary V. Vaughan
+]]
 --[[--
  Additional commands useful for writing command-line specs.