+++ lib/specl/badargs.lua
@@ -1,22 +1,7 @@
--- Bad argument diagnosis helpers.
--- Written by Gary V. Vaughan, 2014
---
--- Copyright (c) 2014-2016 Gary V. Vaughan
---
--- This program is free software; you can redistribute it and/or modify it
--- under the terms of the GNU General Public License as published by the Free
--- Software Foundation; either version 3, or (at your option) any later
--- version.
---
--- This program is distributed in the hope that it will be useful, but WITHOUT
--- ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
--- FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
--- more details.
---
--- You should have received a copy of the GNU General Public License along with
--- this program; see the file COPYING.  If not, a copy can be downloaded from
--- <http://www.gnu.org/licenses/gpl.html>.
-
+--[[
+  Bad argument diagnosis helpers.
+  Copyright (C) 2014-2016, 2018 Gary V. Vaughan
+]]
 --[[--
  Additional commands useful for writing API argument specs.