+++ lib/specl/compat.lua
@@ -1,27 +1,7 @@
--- Compatibility between 5.1, 5.2 and 5.3
--- Written by Gary V. Vaughan, 2013
---
--- Copyright (c) 2013-2016 Gary V. Vaughan
---
--- This program is free software; you can redistribute it and/or modify
--- it under the terms of the GNU General Public License as published by
--- the Free Software Foundation; either version 3, or (at your option)
--- any later version.
---
--- This program is distributed in the hope that it will be useful, but
--- WITHOUT ANY WARRANTY; without even the implied warranty of
--- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
--- General Public License for more details.
---
--- You should have received a copy of the GNU General Public License
--- along with this program; see the file COPYING.  If not, write to the
--- Free Software Foundation, Fifth Floor, 51 Franklin Street, Boston,
--- MA 02111-1301, USA.
-
-
--- Lua 5.3 has renamed package.loaders to package.searchers, so set a
--- metatable to make sure assignments and references go to the right
--- one.
+--[[
+  Compatibility between 5.1, 5.2, 5.3 and 5.4
+  Copyright (C) 2014-2016, 2018 Gary V. Vaughan
+]]
 local package_mt = {
   -- These methods only trigger when the referenced key is missing,
   -- so when the client code references the wrong entry, the methods