+++ lib/specl/badargs.lua
@@ -1,5 +1,5 @@
 --[[
- Behaviour Driven Development for Lua 5.1, 5.2 & 5.3.
+ Behaviour Driven Development for Lua 5.1, 5.2, 5.3 & 5.4
  Copyright (C) 2014-2023 Gary V. Vaughan
 ]]
 --[[--