+++ lib/specl/formatter/tap.lua
@@ -1,6 +1,6 @@
 --[[
- Behaviour Driven Development for Lua 5.1, 5.2, 5.3 & 5.4
- Copyright (C) 2013-2023 Francois Perrad
+ Behaviour Driven Development for Lua 5.1, 5.2, 5.3, 5.4 & 5.5
+ Copyright (C) 2013-2026 Francois Perrad
 ]]
 
 -- Test Anything Protocol style formatter.